Avoid content reading in `TD.toInputMessageContent` when passing content URIs
This improves time between showing the sending file in a chat, but has a drawback that some files will be sent as regular documents (not as videos/GIF/audio). This is up to third-parties to fix (e.g. when selecting a file from Google Drive, none of `MediaStore.PickerMediaColumns` are filled, which are essential to properly display the selected content).